requirements.txt
selenium>=4.0.0
undetected-chromedriver>=3.5.0
requests>=2.31.0
python-dotenv>=1.0.0 